+++ cmd/juggler/server/worker_module.go
@@ -23,6 +23,7 @@ var workerSDKImports = map[string]string{
 	"juggler/strategy-type":          "/sdk/strategy-type.js",
 	"juggler/command-type":           "/sdk/command-type.js",
 	"juggler/info-card-type":         "/sdk/info-card-type.js",
+	"juggler/pinboard":               "/sdk/pinboard.js",
 	"juggler/pinboard-item-type":     "/sdk/pinboard-item-type.js",
 	"juggler/file-viewer":            "/sdk/file-viewer.js",
 	"juggler/file-source":            "/sdk/file-source.js",